Create a ContactForm component based on the contents of ContentSection. Write a small script that renders the contents of ContactForm on load. Modify gulp webpack task to create bundle. TODO: use webpack flags to improve size of client bundle
